feat: 프로젝트 스캐폴딩 & 의존성 설정 (#387)
