+++ .worktrees/task-2fdac7b2/tests/conftest.py
@@ -1,5 +1,5 @@
 import asyncio
-from collections.abc import AsyncGenerator, Generator
+from collections.abc import AsyncGenerator
 
 import pytest
 from httpx import ASGITransport, AsyncClient
@@ -8,7 +8,7 @@ from app.main import app
 
 
 @pytest.fixture(scope="session")
-def event_loop() -> Generator[asyncio.AbstractEventLoop, None, None]:
+def event_loop():
     loop = asyncio.new_event_loop()
     yield loop
     loop.close()
