+++ .worktrees/task-2fdac7b2/tests/test_main.py
@@ -1,0 +1,19 @@
+import pytest
+from httpx import AsyncClient
+
+
+@pytest.mark.asyncio
+async def test_root(client: AsyncClient) -> None:
+    response = await client.get("/")
+    assert response.status_code == 200
+    data = response.json()
+    assert data["message"] == "Personal Jira API"
+
+
+@pytest.mark.asyncio
+async def test_health_check(client: AsyncClient) -> None:
+    response = await client.get("/health")
+    assert response.status_code == 200
+    data = response.json()
+    assert data["status"] == "healthy"
+    assert "app" in data
