--- .worktrees/task-3a78ad64/.env
@@ -1,4 +1,0 @@
-DATABASE_URL=sqlite+aiosqlite:///./dev.db
-APP_NAME=fastapi-app
-DEBUG=true
-DB_ECHO=true
