--- .worktrees/task-3a78ad64/.env.example
@@ -1,5 +1,0 @@
-DATABASE_URL=postgresql+asyncpg://postgres:postgres@localhost:5432/app_db
-APP_NAME=fastapi-app
-DEBUG=false
-API_V1_PREFIX=/api/v1
-DB_ECHO=false
