+++ docs/work-log.md
@@ -15,3 +15,14 @@
 - **Summary**: FastAPI 프로젝트 구조 생성 완료. TDD 원칙에 따라 tests/(conftest, test_config, test_main, test_database) → app/(config, database, main) 순서로 구성. pyproject.toml에 전체 의존성 정의, pydantic-settings 기반 Settings 클래스, asyncio 기
 - **Files**: 13개
 
+### [2026-03-24 04:26:16] Epic Committed
+- **Epic**: Personal Jira 백엔드 API
+- **Issues**: 10개 생성
+
+- [2026-03-24 04:27:31] **agent-backend** → busy (task: 2fdac7b2-f169-4313-a6ba-16e2034d3eec)
+### [2026-03-24 04:28:57] Review Request — SUCCESS
+- **Agent**: agent-backend
+- **Task**: 2fdac7b2-f169-4313-a6ba-16e2034d3eec
+- **Summary**: 프로젝트 스캐폴딩 완료: app/(main.py, config.py, __init__.py), app/routers/, tests/, alembic/ 디렉토리 구조 생성. pyproject.toml에 FastAPI/SQLAlchemy/Alembic 등 의존성 정의. FastAPI 인스턴스(title='Personal Jira') + lifespan 핸들러 
+- **Files**: 7개
+
